Minor Improvements
Minor improvements with the files, shortening of colour codes and removal of all "0px" instances from the CSS file, removing excess whitespace from .html and .js files.

Added a link to Roland Rytz's Atlas Map to the author division where previously it was just href="/" which took you to the root directory of the site.