Updated descriptions for MLP artwork
As I was part of the main force behind putting up MLP artwork during
r/place, I decided to update the descriptions in order to better
description these artworks themselves and the story behind putting
them on r/place. The artworks that I updated descriptions for
include:
   id		   name
 * id: 000367- Rainbow Dash
 * id: two6gx- My Little Pony Crystal Mural
 * id: twqrbr- Giant Derpy
 * id: twvnpr- The Ride Never Ends
 * id: twxl8p- MLP Raid Counter

Note that the names of the entries might be different from the current
ones shown on the Github. As a result, I also included the ids.