Update Berserk MLP entry to sound less aggressive
While the MLP fandomw was initially annoyed by the inclusion of the
Guts picture to the Berserk mural due to feeling betrayed by the
Berserk Pixel discord, the fandom did eventually the
acceptance of the piece due to its high quality and being
a respectful tribute. I tried to make this more clear in the
description of the art.