Update Unturned Entries
I updated each of these entries based on GreatHeroJ's thread on the
r/placeAtlas2 discord.

Note that I put GreatHeroJ's name as the submitter as they were
the one that had given me these recommendations in the first place.
As a result, their name might have to be changed later on in order
match their Reddit name.