Cleaned up calculateCenter function
While I still don't understand what it's doing exactly, I noticed that
the 2 iterations of the loop are unnecessary, and afterwards found out
that the change to area is equal to f. Furthermore I replaced the `~~`
flooring with Math.floor for better readability, since this function
get's executed once per submission and the speed benefits of `~~` are
neglectable.